Cisco CSCuj13135: Data corruption in  MPI_Bsend_ator_c
Do not assume that the "size" passed to alloc_send() will be the same as
the size of the message the resulting fragment will hold when
usnic_send() is called.  This means usnic_send()/usnic_put() can never
trust any pre-computed size values, and are only allowed to look at the
lengths and pointers of the elements in the desc SG list.

This commit was SVN r29183.

+++ ompi/mca/btl/usnic/btl_usnic_recv.c
@@ -283,7 +283,7 @@ opal_output(0, "Start PUT to %p\n", chunk_hdr->ch_hdr.put_addr);
 #if MSGDEBUG2
                 opal_output(0, "  large FRAG complete, pass up %p, %"PRIu64" bytes, tag=%d\n",
                         desc.des_dst->seg_addr.pval, desc.des_dst->seg_len,
-                        chunk_hdr->ch_hdr.tag);
+                        (int)chunk_hdr->ch_hdr.tag);
 #endif
                 reg = mca_btl_base_active_message_trigger +
                     chunk_hdr->ch_hdr.tag;
